changed id type

+++ src/main/java/com/apautomation/model/Employee.java
@@ -13,15 +13,15 @@ public class Employee {
 	
 	@Id
 	@GeneratedValue
-	private int id;
+	private Long id;
 	private String username;
 	private String name;
 	private String email;
 	
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getUsername() {
